# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Discord Music Bot API
  version: 1.0.0
  description: |
    RESTful API for the Discord Music Bot system.

    Provides endpoints for:
    - Analytics and metrics
    - Guild management
    - Music queue operations
    - Premium subscriptions
    - Webhook integrations (Stripe, Discord)

    ## Authentication
    All endpoints require API key authentication via the `X-API-Key` header,
    except for webhook endpoints which use signature verification.

    ## Rate Limiting
    - General endpoints: 100 requests/minute
    - Search endpoints: 10 requests/minute
    - Webhook endpoints: 50 requests/minute

  contact:
    name: Discord Bot Team
    url: https://github.com/your-org/discord-bot
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.discordmusicbot.com/api/v1
    description: Production server (global)
  - url: https://staging-api.discordmusicbot.com/api/v1
    description: Pre-production / staging server
  - url: https://sandbox-api.discordmusicbot.com/api/v1
    description: Internal development sandbox (VPN only)
...
security:
  - ApiKeyAuth: []
paths:
...
  /analytics/guilds:
    get:
      tags:
        - Analytics
      summary: Get guild statistics
      description: Returns statistics for all guilds
      operationId: getGuildAnalytics
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Guild analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuildAnalyticsList'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /guilds:
    get:
      tags:
        - Guilds
      summary: List guilds connected to the bot
      description: Returns a paginated list of guilds synchronized from Discord along with their availability status.
      operationId: listGuilds
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Guild list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuildList'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /guilds/{guildId}:
    get:
      tags:
        - Guilds
      summary: Get guild information
      description: Returns information about a specific guild
      operationId: getGuild
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      responses:
        '200':
          description: Guild retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guild'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /guilds/{guildId}/settings:
    get:
      tags:
        - Guilds
      summary: Get guild settings
      description: Returns settings for a specific guild
      operationId: getGuildSettings
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      responses:
        '200':
          description: Settings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuildSettings'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags:
        - Guilds
      summary: Update guild settings
      description: Updates settings for a specific guild
      operationId: updateGuildSettings
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuildSettingsUpdate'
      responses:
        '200':
          description: Settings updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuildSettings'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /player/{guildId}/now-playing:
    get:
      tags:
        - Music
      summary: Get live playback state
      description: Returns the current playback state reported by the audio service cache
      operationId: getNowPlayingState
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      responses:
        '200':
          description: Playback state retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NowPlaying'
        '404':
          description: Guild not found or no playback data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /player/{guildId}/events:
    get:
      tags:
        - Music
      summary: Subscribe to live playback events
      description: |-
        Server-Sent Events (SSE) stream that emits `snapshot` and `update` payloads when the audio service publishes
        `discord-bot:ui:now`. Clients should include the API key via query string (`?apiKey=`) and keep the connection alive.
      operationId: subscribePlayerEvents
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      responses:
        '200':
          description: Event stream opened
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /player/{guildId}/controls:
    post:
      tags:
        - Music
      summary: Send remote playback action
      description: Issues a playback action (toggle, skip, stop, shuffle, mute, volume) against the active guild player
      operationId: sendPlayerControl
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerControlRequest'
      responses:
        '200':
          description: Command accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerControlResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
...
  /search:
    get:
      tags:
        - Search
      summary: Search for music
      description: Searches for music across multiple platforms
      operationId: searchMusic
      parameters:
        - name: query
          in: query
          description: Search query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - name: source
          in: query
          description: Search source
          required: false
          schema:
            type: string
            enum: [youtube, spotify, soundcloud, all]
            default: all
        - name: limit
          in: query
          description: Maximum number of results
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 25
            default: 10
      responses:
        '200':
          description: Search results retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '429':
          $ref: '#/components/responses/RateLimitError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    GuildIdParam:
      name: guildId
      in: path
      description: Discord guild ID
      required: true
      schema:
        type: string
        pattern: '^\d{17,19}$'
    PageParam:
      name: page
      in: query
      description: Page number
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequestError:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      headers:
        X-RateLimit-Limit:
          description: Request limit per time window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Unix timestamp when limit resets
          schema:
            type: integer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
        timestamp:
          type: string
          format: date-time
...
    GuildAnalyticsList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/GuildAnalytics'
        pagination:
          $ref: '#/components/schemas/Pagination'
    GuildAnalytics:
      type: object
      properties:
        guildId:
          type: string
        tracksPlayed:
          type: integer
        commandsExecuted:
          type: integer
        activeUsers:
          type: integer
        averageSessionDuration:
          type: integer
        tier:
          type: string
          enum: [FREE, TRIAL, PREMIUM]
    Guild:
      type: object
      properties:
        id:
          type: string
          description: Discord guild ID
        name:
          type: string
          description: Friendly guild name pulled from Discord
        icon:
          type: string
          format: uri
          nullable: true
          description: CDN URL for the guild icon, if available
        memberCount:
          type: integer
          nullable: true
        available:
          type: boolean
          description: Indicates whether the guild is currently reachable by the bot
        subscriptionTier:
          type: string
          description: Plan tier configured in ServerConfiguration
        isPremium:
          type: boolean
          description: True when the guild can access premium-only controls
    GuildList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Guild'
        pagination:
          $ref: '#/components/schemas/Pagination'
        timestamp:
          type: string
          format: date-time
...
    GuildSettings:
      type: object
      properties:
        guildId:
          type: string
        volume:
          type: integer
          minimum: 0
          maximum: 200
        autoplay:
          type: boolean
        autoplayMode:
          type: string
          enum: [similar, artist, genre, mixed]
        loopMode:
          type: string
          enum: [off, track, queue]
        streamable:
          type: boolean
          description: Indicates whether the API can proxy the audio for web playback
        djRoleId:
          type: string
          nullable: true
        notificationChannelId:
          type: string
          nullable: true
    GuildSettingsUpdate:
      type: object
      properties:
        volume:
          type: integer
          minimum: 0
          maximum: 200
        autoplay:
          type: boolean
        autoplayMode:
          type: string
          enum: [similar, artist, genre, mixed]
        loopMode:
          type: string
          enum: [off, track, queue]
        djRoleId:
          type: string
          nullable: true
        notificationChannelId:
          type: string
          nullable: true
...
    PlayerControlRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: [toggle, skip, stop, shuffle, clear, previous, mute, volume]
        value:
          type: integer
          minimum: 0
          maximum: 200
          description: Required when action is volume
    PlayerControlResponse:
      type: object
      properties:
        accepted:
          type: boolean
...
    NowPlaying:
      type: object
      properties:
        guildId:
          type: string
        title:
          type: string
        author:
          type: string
          nullable: true
        uri:
          type: string
          nullable: true
        artworkUrl:
          type: string
          nullable: true
        durationMs:
          type: integer
        positionMs:
          type: integer
        isStream:
          type: boolean
        paused:
          type: boolean
        repeatMode:
          type: string
          enum: [off, track, queue]
        queueLen:
          type: integer
        hasTrack:
          type: boolean
        canSeek:
          type: boolean
        volume:
          type: integer
          minimum: 0
          maximum: 200
        autoplay:
          type: boolean
        autoplayMode:
          type: string
          enum: [off, similar, artist, genre, mixed]
        filter:
          type: object
          properties:
            id:
              type: string
            label:
              type: string
            description:
              type: string
          nullable: true
        updatedAt:
          type: integer
          format: int64
          description: Epoch milliseconds when the cache entry was produced
    SearchResults:
      type: object
      properties:
        query:
          type: string
        source:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        total:
          type: integer
    SearchResult:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        url:
          type: string
        duration:
          type: integer
        thumbnail:
          type: string
        source:
          type: string
          enum: [youtube, spotify, soundcloud]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean